% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_ci}
\alias{add_ci}
\title{Adding confidence intervals to Nonparametric kernel smoothing (NKS)}
\usage{
add_ci(x, y, h, beta)
}
\arguments{
\item{x}{x serves as the input: x coordinate}

\item{y}{y serves as the response: y coordinate}

\item{h}{h serves as the bandwidth used for local fitting}

\item{beta}{beta serves as the values of fitted \eqn{\hat{\theta}(x)} and \eqn{\hat{\theta}^\prime(x)}}
}
\value{
95\% CI's associated with each beta.
}
\description{
This function fits a local quadratic regression model \eqn{y = \theta(x)} to the data using a Gaussian kernel and evaluates equilibrium points of the fitted model \eqn{\hat{\theta}^\prime(x)} at the data points. Finally, estimates and associated CI's are returned
}
\examples{
set.seed(1234)
x <- runif(1000)
y <- runif(1000)
z <- x^2 + y^2
cond_dens <- estim_conditional(x, y, z, f_eval = 0.25)
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
