% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{affine_unif}
\alias{affine_unif}
\title{Affine transformation to [0,1]}
\usage{
affine_unif(x)
}
\arguments{
\item{x}{Input for transformation}
}
\value{
A vector of output values from the affine transformation
}
\description{
This function affine transforms data to lie in compact [0,1].
}
\examples{
## simulation from y = x*x
set.seed(1234)
x <- runif(1000)
y <- data_gen(x, 2)
z <- affine_unif(y)
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
