% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{calc_binom_odds}
\alias{calc_binom_odds}
\title{Calculating binomial odds}
\usage{
calc_binom_odds(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix of binomial odds
}
\description{
Calculating binomial odds
}
