% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS23.R, R/LAS15.R, R/LAS21.R
\name{calc_las}
\alias{calc_las}
\alias{calc_wl_cas23}
\alias{calc_post_tx_cas23}
\alias{calc_wl_las15}
\alias{calc_post_tx_las15}
\alias{calc_wl_las21}
\alias{calc_post_tx_las21}
\title{Calculate LAS/CAS Linear Predictors}
\usage{
calc_wl_cas23(
  data = NULL,
  c_id = c_id,
  dx_grp = dx_grp,
  dx = dx,
  age = age,
  bmi = bmi,
  funstat = funstat,
  walk6m = walk6m,
  o2rest = o2rest,
  pap_syst = pap_syst,
  pap_mean = pap_mean,
  pco2 = pco2,
  pco2_15 = pco2_15,
  cont_mech = cont_mech,
  creat = creat,
  bili = bili,
  ecmo = ecmo
)

calc_post_tx_cas23(
  data = NULL,
  c_id = c_id,
  dx_grp = dx_grp,
  dx = dx,
  age = age,
  funstat = funstat,
  pap_mean = pap_mean,
  walk6m = walk6m,
  cont_mech = cont_mech,
  creat = creat,
  ci = ci,
  ecmo = ecmo
)

calc_wl_las15(
  data = NULL,
  c_id = c_id,
  dx_grp = dx_grp,
  dx = dx,
  age = age,
  bmi = bmi,
  funstat = funstat,
  walk6m = walk6m,
  o2rest = o2rest,
  pap_syst = pap_syst,
  pap_mean = pap_mean,
  pco2_15 = pco2_15,
  pco2 = pco2,
  cont_mech = cont_mech,
  creat = creat,
  bili = bili,
  cvp = cvp,
  bili_50 = bili_50,
  fvc = fvc,
  ci = ci,
  diab = diab
)

calc_post_tx_las15(
  data = NULL,
  c_id = c_id,
  dx_grp = dx_grp,
  dx = dx,
  age = age,
  o2rest = o2rest,
  pap_mean = pap_mean,
  walk6m = walk6m,
  cont_mech = cont_mech,
  creat = creat,
  ci = ci,
  funstat = funstat,
  creat_150 = creat_150
)

calc_wl_las21(
  data = NULL,
  c_id = c_id,
  dx_grp = dx_grp,
  dx = dx,
  age = age,
  bmi = bmi,
  funstat = funstat,
  walk6m = walk6m,
  o2rest = o2rest,
  pap_syst = pap_syst,
  pap_mean = pap_mean,
  pco2_15 = pco2_15,
  pco2 = pco2,
  cont_mech = cont_mech,
  creat = creat,
  bili = bili
)

calc_post_tx_las21(
  data = NULL,
  c_id = c_id,
  dx_grp = dx_grp,
  dx = dx,
  age = age,
  o2rest = o2rest,
  pap_mean = pap_mean,
  walk6m = walk6m,
  cont_mech = cont_mech,
  creat = creat,
  ci = ci
)
}
\arguments{
\item{data}{dataset of patient data}

\item{c_id}{candidate id}

\item{dx_grp}{diagnosis group (A, B, C, or D)}

\item{dx}{diagnosis (numeric, see details for ones in LAS/CAS)}

\item{age}{age in years}

\item{bmi}{bmi {\out{(kg/m<sup>2</sup>)}}}

\item{funstat}{functional status (1 = No assistance needed, 2 = some assistance needed, 3 = total assistance needed)}

\item{walk6m}{six minute walk distance (ft)}

\item{o2rest}{amount of oxygen received if needed while resting (L/min)}

\item{pap_syst}{systolic pulmonary artery pressure (mmHg)}

\item{pap_mean}{mean pulmonary artery pressure (mmHg)}

\item{pco2}{\out{pco<sub>2</sub> (mmHG)}}

\item{pco2_15}{\out{pco<sub>2</sub> increase by at least 15\% threshold (0 or 1)}}

\item{cont_mech}{continuous mechangical ventilation (0 or 1)}

\item{creat}{creatinine (mg/dL)}

\item{bili}{bilirubin (mg/dL)}

\item{ecmo}{ECMO (0 or 1)}

\item{ci}{cardiac index}

\item{cvp}{central venous pressure (mmHg)}

\item{bili_50}{bilirubin increase of 50\% (0 or 1)}

\item{fvc}{forced vital capapcity (\%)}

\item{diab}{diabetes (0 or 1)}

\item{creat_150}{creatine increase >150\% (0 or 1)}
}
\value{
a dataset with c_id and linear predictor for waitlist or post-transplant linear predictor according to LAS/CAS equations
}
\description{
Calculate LAS/CAS Linear Predictors
}
\details{
Numeric Diagnosis Options
\itemize{
\item 214,1608: bronchiestais
\item 1605:  sarcoidosis
\item 1519,1613: pulmonary fibrosis (non-idiopathetic),
\item 1611: lymphangioleiomyomatosis
\item 1612: Obliterative bronchiolitis
\item 1500,1501,1502,1548,1549: eisenmenger's
}
}
\examples{
c0 <- tibble(c_id = as.integer(c(1, 2, 3)), center = factor(c('12', '54', '4'), levels = c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', '25', '26', '27', '28', '29', '30', '31', '32', '33', '34', '35', '36', '37', '38', '39', '40', '41', '42', '43', '44', '45', '46', '47', '48', '49', '50', '51', '52', '53', '54', '55', '56', '57', '58', '59', '60', '61')),
listing_day = as.integer(c(2208, 984, 1564)), male = as.integer(c(1, 0, 0)), dx_grp = c('A', 'D', 'D'), race_eth = factor(c('NH Black', 'NH White', 'NH Black'), levels = c('NH White', 'NH Black', 'Hispanic', 'Asian', 'Amer. Ind', 'Hawaiian', 'Multi/Other')), age_at_listing = as.numeric(c(57.5013698630137, 47.3315068493151, 60.4246575342466)), airway = as.numeric(c(-0.897327510985039, 0.541481421974704, -1.01115201702156)), oxygen = as.numeric(c(0.167831951570743, 2.48090561590646, 0.710183343361671)), abo = c('O', 'A', 'A'), hgt_in = as.numeric(c(70, 60.5, 65)), hgt_cm = as.numeric(c(177.8, 153.67, 165.1)), wgt_kg = as.numeric(c(69.46, 68.87, 73.52)), bmi = as.numeric(c(21.972084760496, 29.1643230646614, 26.9718882633031)), resp_supp = factor(c('2', '2', '1'), levels = c('1', '2', '3', '4')),
surg_type = factor(c('E', 'D', 'S'), levels = c('D', 'E', 'S')), diab = as.integer(c(1, 0, 1)), fev1 = as.numeric(c(12, 58, 21)), fvc = as.numeric(c(51, 74, 14)), pco2 = as.numeric(c(72, 39, 75)), pf = as.numeric(c(308.133653156936, 849.217897083732, 388.523032567967)), po2 = as.numeric(c(146, 184, 102)), pap_mean = as.numeric(c(38.6666666666667, 17, 34.3333333333333)), o2_freq = as.numeric(c(1, 1, 3)), vent = factor(c('None', 'None', 'None'), levels = c('BiPap', 'Mechanical', 'None', 'CPAP')), walk6m = as.numeric(c(810, 977, 877)), o2 = as.integer(c(2, 2, 7)), bili = as.numeric(c(0.36, 0.25, 0.56)),
creat = as.numeric(c(0.69, 1.25, 0.77)), pap_syst = as.numeric(c(56, 28, 55)), ci = as.numeric(c(2.54, 1.81, 3.14)), funstat = as.numeric(c(2, 2, 1)), ecmo = as.numeric(c(0, 0, 0)), cvp = as.numeric(c(14, 6, 4)), pco2_15 = as.numeric(c(0, 0, 1)), pra = as.numeric(c(0, 0, 0)), pld = as.numeric(c(0, 0, 0)), creat_150 = as.numeric(c(0, 0, 0)), bili_50 = as.numeric(c(0, 0, 0)), age = as.numeric(c(57.5013698630137, 47.3315068493151, 60.4246575342466)), hgt_cm1 = as.numeric(c(178, 154, 165)), o2rest = as.numeric(c(2, 2, 0)), cont_mech = as.numeric(c(0, 0, 0)), dx = as.numeric(c(1608, -99, 1613)), pabo = as.numeric(c(0.500181752090149, 0.13958560523446, 0.13958560523446)), phgt = as.numeric(c(0.363702359346643, 0.489655172413793, 0.110344827586207)), pcpra = as.numeric(c(0, 0, 0)))
calc_wl_cas23(c0)
calc_post_tx_cas23(c0)
calc_wl_las21(c0)
calc_post_tx_las21(c0)
calc_wl_las15(c0)
calc_post_tx_las15(c0)
#'c0 <- tibble(c_id = as.integer(c(1, 2, 3)), center = factor(c('12', '54', '4'), levels = c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', '25', '26', '27', '28', '29', '30', '31', '32', '33', '34', '35', '36', '37', '38', '39', '40', '41', '42', '43', '44', '45', '46', '47', '48', '49', '50', '51', '52', '53', '54', '55', '56', '57', '58', '59', '60', '61')),
listing_day = as.integer(c(2208, 984, 1564)), male = as.integer(c(1, 0, 0)), dx_grp = c('A', 'D', 'D'), race_eth = factor(c('NH Black', 'NH White', 'NH Black'), levels = c('NH White', 'NH Black', 'Hispanic', 'Asian', 'Amer. Ind', 'Hawaiian', 'Multi/Other')), age_at_listing = as.numeric(c(57.5013698630137, 47.3315068493151, 60.4246575342466)), airway = as.numeric(c(-0.897327510985039, 0.541481421974704, -1.01115201702156)), oxygen = as.numeric(c(0.167831951570743, 2.48090561590646, 0.710183343361671)), abo = c('O', 'A', 'A'), hgt_in = as.numeric(c(70, 60.5, 65)), hgt_cm = as.numeric(c(177.8, 153.67, 165.1)), wgt_kg = as.numeric(c(69.46, 68.87, 73.52)), bmi = as.numeric(c(21.972084760496, 29.1643230646614, 26.9718882633031)), resp_supp = factor(c('2', '2', '1'), levels = c('1', '2', '3', '4')),
surg_type = factor(c('E', 'D', 'S'), levels = c('D', 'E', 'S')), diab = as.integer(c(1, 0, 1)), fev1 = as.numeric(c(12, 58, 21)), fvc = as.numeric(c(51, 74, 14)), pco2 = as.numeric(c(72, 39, 75)), pf = as.numeric(c(308.133653156936, 849.217897083732, 388.523032567967)), po2 = as.numeric(c(146, 184, 102)), pap_mean = as.numeric(c(38.6666666666667, 17, 34.3333333333333)), o2_freq = as.numeric(c(1, 1, 3)), vent = factor(c('None', 'None', 'None'), levels = c('BiPap', 'Mechanical', 'None', 'CPAP')), walk6m = as.numeric(c(810, 977, 877)), o2 = as.integer(c(2, 2, 7)), bili = as.numeric(c(0.36, 0.25, 0.56)),
creat = as.numeric(c(0.69, 1.25, 0.77)), pap_syst = as.numeric(c(56, 28, 55)), ci = as.numeric(c(2.54, 1.81, 3.14)), funstat = as.numeric(c(2, 2, 1)), ecmo = as.numeric(c(0, 0, 0)), cvp = as.numeric(c(14, 6, 4)), pco2_15 = as.numeric(c(0, 0, 1)), pra = as.numeric(c(0, 0, 0)), pld = as.numeric(c(0, 0, 0)), creat_150 = as.numeric(c(0, 0, 0)), bili_50 = as.numeric(c(0, 0, 0)), age = as.numeric(c(57.5013698630137, 47.3315068493151, 60.4246575342466)), hgt_cm1 = as.numeric(c(178, 154, 165)), o2rest = as.numeric(c(2, 2, 0)), cont_mech = as.numeric(c(0, 0, 0)), dx = as.numeric(c(1608, -99, 1613)), pabo = as.numeric(c(0.500181752090149, 0.13958560523446, 0.13958560523446)), phgt = as.numeric(c(0.363702359346643, 0.489655172413793, 0.110344827586207)), pcpra = as.numeric(c(0, 0, 0)))
calc_wl_cas15(c0)
calc_post_tx_cas15(c0)

}
