% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{calc_multi_odds}
\alias{calc_multi_odds}
\title{Calculating multinomial odds}
\usage{
calc_multi_odds(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix of multinomial odds
}
\description{
Calculating multinomial odds
}
