% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacif.R
\name{calccinc}
\alias{calccinc}
\title{CIF computation}
\usage{
calccinc(events, cause, N, ptime, timepoint)
}
\description{
calccinc computes CIF, log ratio of CIFs, and their standard deviation (using delta method derivation).
}
\seealso{
metacinc
}
