% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cas.R
\name{calculate_sub_cas}
\alias{calculate_sub_cas}
\alias{calculate_cas}
\alias{calculate_cas_dist}
\title{Calculate Composite Allocation Score (CAS)}
\usage{
calculate_sub_cas(
  data,
  wl_model = "CAS23",
  post_tx_model = "CAS23",
  wl_weight = NA,
  post_tx_weight = NA,
  bio_weight = NA,
  peds_weight = NA,
  pld_weight = NA,
  wl_cap = NA,
  post_tx_cap = NA,
  abo_weight = NA,
  height_weight = NA,
  cpra_weight = NA,
  checks = TRUE
)

calculate_cas_dist(
  match_data,
  efficiency_weight = NA,
  cost_weight = NA,
  distance_weight = NA,
  checks = TRUE
)
}
\arguments{
\item{data}{dataset of patients containing covaraites for las calculations and other continuous distribution factors}

\item{wl_model}{model to use in waitlist component of the CAS}

\item{post_tx_model}{model to use in the post-transplant component of the CAS}

\item{wl_weight}{weight applied to waitlist model}

\item{post_tx_weight}{weight applied to post transplant model}

\item{bio_weight}{weight applied to biologoical measures (height, abo (blood), and hla)}

\item{peds_weight}{weight applied to give a bonus to pediatric patients}

\item{pld_weight}{weight given as bonus if a prior living donor}

\item{wl_cap}{number of days to cap the waitlist component CAS score at CAS}

\item{post_tx_cap}{number of days to cap the post transplant component of the CAS}

\item{abo_weight}{weight for blood type compatability (if not specified will default to 1/3 of bio weight)}

\item{height_weight}{weight for height compatability (if not specified will default to 1/3 of bio weight)}

\item{cpra_weight}{weight for cpra compatability (if not specified will default to 1/3 of bio weight)}

\item{checks}{whether or not to check the conditions and display warnings, this is there to not check conditions every time the simulation is iterated}

\item{match_data}{dataset of patients containing covaraites for las calculations and other continuous distribution factors}

\item{efficiency_weight}{weight given for efficiency (cost and distance)}

\item{cost_weight}{weight given for cost part of CAS (if not specified is half of efficiency weight)}

\item{distance_weight}{weight given for distance part of CAS (if not specified is half of efficiency weight)}
}
\value{
\code{calculate_sub_cas} a dataset of candidate, CAS score on a 0 to 1 scale (lu_score) without the efficiency weight,
and ov_rank which is the overall rank of each candidate. 1 is the highest CAS score
\code{calculate_cas_dist} a dataset of matches with lu_score that contains the efficiency weight. It will contain the estimaed cost (pcost) for efficiency calculation.

a dataset complete with the efficient/distance metric
}
\description{
Calcualte the subCAS
}
\examples{
calculate_sub_cas(data = syn_cands, wl_model = "CAS23", post_tx_model = "CAS23",
 wl_weight = .25, post_tx_weight = .25, bio_weight = .15, peds_weight = .2,
 pld_weight = 0.05, wl_cap = 365, post_tx_cap = 1825)

calculate_cas_dist(syn_matches, efficiency_weight = 0.1)
}
