% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cas.R
\name{calculate_cas_dist}
\alias{calculate_cas_dist}
\title{Calculate continuous distribution}
\usage{
calculate_cas_dist(
  match_data,
  efficiency_weight = NA,
  cost_weight = NA,
  distance_weight = NA,
  checks = TRUE
)
}
\arguments{
\item{match_data}{dataset of patients containing covaraites for las calculations and other continuous distribution factors}

\item{efficiency_weight}{weight given for efficiency (cost and distance)}

\item{cost_weight}{weight given for cost part of CAS (if not specified is half of efficiency weight)}

\item{distance_weight}{weight given for distance part of CAS (if not specified is half of efficiency weight)}

\item{checks}{whether or not to check the conditions and display warnings, this is there to not check conditions every time the simulation is iterated}
}
\value{
a dataset complete with the efficient/distance metric
}
\description{
Calculate continuous distribution
}
