% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cas.R
\name{calculate_cas_nodist}
\alias{calculate_cas_nodist}
\title{Calculate continuous distribution}
\usage{
calculate_cas_nodist(
  data,
  wl_model = "CAS23",
  post_tx_model = "CAS23",
  wl_weight = NA,
  post_tx_weight = NA,
  bio_weight = NA,
  peds_weight = NA,
  pld_weight = NA,
  wl_cap = NA,
  post_tx_cap = NA,
  abo_weight = NA,
  height_weight = NA,
  cpra_weight = NA
)
}
\arguments{
\item{data}{dataset of patients containing covaraites for las calculations and other continuous distribution factors}

\item{wl_weight}{weight applied to waitlist model}

\item{post_tx_weight}{weight applied to post transplant model}

\item{bio_weight}{weight applied to biologoical measures (height, abo (blood), and hla)}

\item{peds_weight}{weight applied to give a bonus to pediatric patients}

\item{pld_weight}{weight given as bonus if a prior living donor}

\item{abo_weight}{weight for blood type compatability (if not specified will default to 1/3 of bio weight)}

\item{height_weight}{weight for height compatability (if not specified will default to 1/3 of bio weight)}

\item{cpra_weight}{weight for cpra compatability (if not specified will default to 1/3 of bio weight)}
}
\value{
a dataset of cas calcualted metrics without the efficiency/distance metric
}
\description{
Calculate continuous distribution
}
\examples{
## will add real examples later
## This needs to be thought out, only the distance changes with the donors
cands <- gen_and_spawn_candidates(days = 10)
cas_rank <- calculate_cas_nodist(cands, wl_model = "CAS23", post_tx_model = "CAS23", wl_weight = 0.25, post_tx_weight = 0.25, bio_weight = 0.15, pld_weight = 0.05, peds_weight = 0.2)
}
