% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{calculate_conf_intervals}
\alias{calculate_conf_intervals}
\title{This function calculates confidence intervals for every sample over 10
runs}
\usage{
calculate_conf_intervals(data.inputs)
}
\arguments{
\item{data.inputs}{the input datasheet stored in a csv file in the tables dir}
}
\value{
nothing, saves results within the Confidence_Interval_Calculations dir
}
\description{
This function calculates confidence intervals for every sample over 10
runs
}
