% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS.R
\name{calculate_las}
\alias{calculate_las}
\title{Calcualate LAS scores}
\usage{
calculate_las(
  wl_data,
  wl_model = NULL,
  post_tx_model = NULL,
  wl_cap = NA,
  post_tx_cap = NA,
  wl_weight = NA,
  post_tx_weight = NA
)
}
\arguments{
\item{wl_data}{data to calculate las}

\item{wl_model}{"LAS15","LAS21", "CAS23" or own coxph model for waitlist  (in days)}

\item{post_tx_model}{"LAS15","LAS21", "CAS23" or own coxph model for post-transplant mortality (in days)}

\item{wl_cap}{cap of waitlist survival}

\item{post_tx_cap}{cap of post-transplant survival}

\item{wl_weight}{relative weight of waitlist mortality}

\item{post_tx_weight}{relative weight of post-transplant mortality}
}
\value{
a dataset of candidates and their LAS score on a 0 to 1 scale
}
\description{
Calcualate LAS scores
}
\examples{
cands <- gen_and_spawn_candidates(days = 100)
las_scores <- calculate_las(cands, wl_model = "LAS21", post_tx_model = "LAS21")
}
