% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS.R
\name{cap_survrates}
\alias{cap_survrates}
\title{Title LAS FUNCTIONS}
\usage{
cap_survrates(surv_data, Days = Days, cap = NA)
}
\arguments{
\item{surv_data}{survival data first columsn labeles Days and Survival}

\item{cap}{if you want to cap the survival at a certain time default is max Days}
}
\value{
dataframe of survival rates, but capped at a certain point
}
\description{
Title LAS FUNCTIONS
}
\examples{
desired_survrates <- all_survrates \%>\%
cap_survrates()
#' desired_survrates_180 <- all_survrates \%>\%
cap_survrates(cap = 180)
}
