% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_policy.R
\name{CASPolicy}
\alias{CASPolicy}
\title{Create a CAS policy}
\usage{
CASPolicy(name = "CAS policy", params = list(), notes = "", autofill = TRUE)
}
\arguments{
\item{name}{Policy name.}

\item{params}{A named list of policy parameters.}

\item{notes}{Optional notes.}

\item{autofill}{Whether to fill missing required keys with NA.}
}
\value{
A CASPolicy object.
}
\description{
Create a CAS policy
}
