% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_experiment.R
\name{COMETExperiment}
\alias{COMETExperiment}
\title{Create a COMET experiment}
\usage{
COMETExperiment(
  simulator,
  n_runs,
  seeds = NULL,
  parallel = FALSE,
  workers = NULL,
  notes = ""
)
}
\arguments{
\item{simulator}{A COMETSimulator object.}

\item{n_runs}{Number of runs.}

\item{seeds}{Optional seed or seed vector.}

\item{parallel}{Whether to run in parallel.}

\item{workers}{Optional number of workers.}

\item{notes}{Optional notes.}
}
\value{
A COMETExperiment object.
}
\description{
Create a COMET experiment
}
