% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_simulator.R
\name{COMETSimulator}
\alias{COMETSimulator}
\title{Create a COMET simulator}
\usage{
COMETSimulator(
  config,
  policy,
  candidate_generator = gen_and_spawn_candidates,
  donor_generator = gen_and_spawn_donors,
  iteration_fn = iteration,
  preflight_fn = data_present,
  notes = ""
)
}
\arguments{
\item{config}{A COMETConfig object.}

\item{policy}{A COMETPolicy object.}

\item{candidate_generator}{Candidate generation function.}

\item{donor_generator}{Donor generation function.}

\item{iteration_fn}{Iteration function.}

\item{preflight_fn}{Optional preflight function.}

\item{notes}{Optional notes.}
}
\value{
A COMETSimulator object.
}
\description{
Create a COMET simulator
}
