% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{concat}
\alias{concat}
\title{Simulation Evaluation Helps}
\usage{
concat(COMET)
}
\arguments{
\item{COMET}{a COMET object}
}
\value{
a complete tibble of all the COMET's list, along with survival information
}
\description{
Simulation Evaluation Helps
}
