% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_config.R
\name{configSummary}
\alias{configSummary}
\title{Summarize a COMET configuration}
\usage{
configSummary(config)
}
\arguments{
\item{config}{A COMETConfig object.}
}
\description{
Summarize a COMET configuration
}
