% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_las.R
\name{create_surv_rates}
\alias{create_surv_rates}
\title{Creating complete survival estimates}
\usage{
create_surv_rates(object, cap = NA, ignore = FALSE)
}
\arguments{
\item{object}{either survift or coxph object}

\item{cap}{if you want survival capped at any timepoint, default is max of data}

\item{ignore}{ignore warnings or not}
}
\value{
a complete dataset of survival estimates and time
}
\description{
expand and fill in the data.frame from start to finish. No more missing days
}
