% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS.R
\name{custom_scale}
\alias{custom_scale}
\title{Custom Scaling}
\usage{
custom_scale(x, min, max)
}
\arguments{
\item{x}{raw value double}

\item{min}{minimum possible value}

\item{max}{maximum possible value of scale}
}
\value{
a value between 0 and 1
}
\description{
Custom Scaling
}
\examples{
## Should return 1
custom_scale(24, 0, 24)
## Should return 0.07038462
custom_scale(0.745, -2, 37)
}
