% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_policy.R
\name{CustomPolicy}
\alias{CustomPolicy}
\title{Create a custom policy}
\usage{
CustomPolicy(name = "Custom policy", params = list(), notes = "")
}
\arguments{
\item{name}{Policy name.}

\item{params}{A named list of policy parameters.}

\item{notes}{Optional notes.}
}
\value{
A CustomPolicy object.
}
\description{
Create a custom policy
}
