% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{cutoff}
\alias{cutoff}
\title{CUTOFF}
\usage{
cutoff(COMET, enroll_date, waitlist_date, post_tx_date)
}
\arguments{
\item{COMET}{a COMET object}

\item{enroll_date}{last date to enroll for simulation}

\item{censor_date}{last date to censor individuals post-transplant}
}
\value{
a tibble with simulation data and calucated survival data
}
\description{
CUTOFF
}
