% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{daily_count_las}
\alias{daily_count_las}
\title{Title}
\usage{
daily_count_las(COMET, LAS_FUN = calculate_las, ...)
}
\arguments{
\item{COMET}{a COMET object}
}
\value{
a tibble of daily counts and mean and median las per COMET
}
\description{
Title
}
