% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_gen}
\alias{data_gen}
\title{Data generation}
\usage{
data_gen(x, type = 1)
}
\arguments{
\item{x}{Input for generative model}

\item{type}{Which generative function should we use?}
}
\value{
A vector of output values from the generative exposure model
}
\description{
This function generates data for the simulation study.
}
\examples{
## simulation from y = x*x
set.seed(1234)
x <- runif(1000)
y <- data_gen(x, 2)
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
