% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{DTW_calculate}
\alias{DTW_calculate}
\title{This function calculates the DTW distance bewteen the inferred trajectories
for every cutoff and the flow cytometry data}
\usage{
DTW_calculate(data.inputs, MET.range)
}
\arguments{
\item{data.inputs}{the input datasheet stored in a csv file in the tables dir}

\item{MET.range}{range of time for MET to take place}
}
\value{
nothing, saves the matrix in the DTW_Matrix dir
}
\description{
This function calculates the DTW distance bewteen the inferred trajectories
for every cutoff and the flow cytometry data
}
