% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{entropy_calc}
\alias{entropy_calc}
\title{Differential entropy calculation}
\usage{
entropy_calc(x)
}
\arguments{
\item{x}{Input density estimates for calculation #'}
}
\value{
Differential entropy estimate
}
\description{
This function calculates differential entropy given density estimates
}
\examples{
set.seed(1234)
x <- runif(1000)
p <- density(x)
entropy_calc(p$y)
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
