% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estim_conditional}
\alias{estim_conditional}
\title{Conditional differential entropy calculation}
\usage{
estim_conditional(x, y, z, f_eval = 0.25)
}
\arguments{
\item{x}{x serves as one input of the generative model}

\item{y}{y serves as the output of the generative model}

\item{z}{z serves as another input of the generative model, will serve as conditioning variable}

\item{f_eval}{Evaluation grid will contain \code{frac} fraction of the data points}
}
\value{
Conditional entropy estimates \eqn{H(X \mid Y = y) - H(Z \mid Y = y)} for values of \code{y} in the evaluation grid
}
\description{
This function evaluates the difference of conditional differential entropy of \eqn{H(X \mid Z = z) - H(Y \mid Z = z)} in the generative exposure model \eqn{Y = g(X, Z)} over a grid of values of \eqn{Z}, where the grid is controlled by number of grid points = \code{f_eval*n}, \code{n} being the sample size of the data.
}
\examples{
set.seed(1234)
x <- runif(1000)
y <- runif(1000)
z <- x^2 + y^2
cond_dens <- estim_conditional(x, y, z, f_eval = 0.25)
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
