% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_simulation}
\alias{eval_simulation}
\title{Title}
\usage{
eval_simulation(COMET, group = NULL)
}
\arguments{
\item{COMET}{a COMET object}

\item{group}{what to group by}
}
\value{
a tibble concatted simulatino
}
\description{
Title
}
