% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_simulation}
\alias{eval_simulation}
\alias{evaluation}
\alias{daily_count_las}
\alias{cutoff}
\alias{concat2}
\title{Evaluation}
\usage{
eval_simulation(
  COMET,
  min_enroll_date = 1,
  max_enroll_date = NA,
  wl_censor_date = NA,
  post_tx_censor_date = NA,
  group = NULL
)

daily_count_las(
  COMET,
  min_date = 1,
  max_date = NA,
  LAS_FUN = calculate_las,
  ...
)

cutoff(COMET, enroll_date, waitlist_date, post_tx_date)

concat2(
  COMET,
  min_enroll_date = 1,
  max_enroll_date = NA,
  wl_censor_date = NA,
  post_tx_censor_date = NA
)
}
\arguments{
\item{COMET}{a COMET object}

\item{min_enroll_date}{date to enroll candidates at default is 1}

\item{max_enroll_date}{date to stop the enrollment period, defaults is maximum date found}

\item{wl_censor_date}{date to censor waitlist outcomes, default is maximum date found}

\item{post_tx_censor_date}{date to censor post-transplant outcomes, default is maximum date found}

\item{group}{summary statistics by a certain group}

\item{min_date}{start date for daily las count}

\item{LAS_FUN}{function to calculate LAS/CAS score}

\item{...}{arguments pased to LAS_FUN}

\item{enroll_date}{last date to enroll for simulation}

\item{max_data}{end date for find daily las_count (default is maximum listing day of COMEt)}

\item{censor_date}{last date to censor individuals post-transplant}
}
\value{
\code{eval_simulation()} returns a tibble of summary statistics

a tibble of daily counts and mean and median las per COMET

a tibble with simulation data and calucated survival data

a tibble of candidates and time information
}
\description{
Evaluation

Title
}
\examples{
\dontrun{
r1 <- run_simulation(days = 20, can_start = 1000, waitlist_update_freq = 1, post_tx_update_freq = 1,
 match_alg = match_las, wl_model = "LAS15", post_tx_model = "LAS15",
 wl_weight = 1, post_tx_weight = 1, wl_cap = 365, post_tx_cap = 365)
eval_simulation(r1)
}
\dontrun{concat2(r1)}
}
