% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS.R
\name{expected_days}
\alias{expected_days}
\title{Title LAS FUNCTIONS}
\usage{
expected_days(data, surv_rates, lp)
}
\arguments{
\item{data}{dataset containing linear predictor and a linear predictor for waitlist}

\item{surv_rates}{vector of survival rates}

\item{lp}{linear predictor from a model}

\item{id}{patient id}
}
\value{
expected mean survival
}
\description{
Title LAS FUNCTIONS
}
\examples{
# all_survrates <- readxl::read_xlsx("/proj/DaltonLab/projects/lungTx/p0059/data_raw/LAS_refit2_betas.xlsx", sheet = 2, skip =2)

desired_survrates <- all_survrates \%>\%
cap_survrates()

df <- data.frame(c_id = 1:20, lp_wl = rnorm(20,0,2))
df2 <- expected_days(df, desired_survrates$Survival, lp = lp_wl)
}
