% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expected_survival}
\alias{expected_survival}
\title{Expected Survival}
\usage{
expected_survival(lp, surv_rate)
}
\arguments{
\item{lp}{vector of linear predictors}

\item{surv_rate}{vector of survival times}
}
\value{
if surv_rate is of length 1 returns probability of survival, if surv_rate is a vector resturns RMST
}
\description{
Calculates Expected Survival Probability or Restricted Mean Survival Time (RMST)
}
\examples{
lps <- c(0, 1, -.5)
## Expected Survival
expected_survival(lps, 0.85)
## Restricted Mean Survival Time
expected_survival(lps, seq(1,.85, -.0005))
}
