% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_experiment.R
\name{experimentMetrics}
\alias{experimentMetrics}
\title{Get per-run metrics from a COMETExperimentResult}
\usage{
experimentMetrics(x)
}
\arguments{
\item{x}{A COMETExperimentResult object.}
}
\description{
Get per-run metrics from a COMETExperimentResult
}
