% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{find.min.alpha}
\alias{find.min.alpha}
\title{This function finds the mse_total given a certain alpha}
\usage{
find.min.alpha(
  alpha,
  E_cad,
  hybrid,
  ZEB,
  M_sc,
  Mu_sc,
  eq,
  ref_eq_day,
  timepoints
)
}
\arguments{
\item{alpha}{parameter alpha}

\item{E_cad}{Epithelial percentage, just named E_cad}

\item{hybrid}{hybrid percentage}

\item{ZEB}{Mesenchymal percentage, just named ZEB}

\item{M_sc}{ratio of M/E at steady state}

\item{Mu_sc}{ratio of H/E at steady state}

\item{eq}{what timepoint to start}

\item{ref_eq_day}{steady state timepoint}

\item{timepoints}{total timepoints}
}
\value{
mse_total
}
\description{
This function finds the mse_total given a certain alpha
}
