% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{find.optimal.cutoff}
\alias{find.optimal.cutoff}
\title{Find the optimal cutoff
Finds the best number of highly variable EMT genes}
\usage{
find.optimal.cutoff(data.input)
}
\arguments{
\item{data.input}{input data to use}
}
\value{
optimal cutoff of highly variable genes
}
\description{
Find the optimal cutoff
Finds the best number of highly variable EMT genes
}
