% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_bw}
\alias{find_bw}
\title{Finding bandwidth for Nonparametric kernel smoothing (NKS)}
\usage{
find_bw(x, y)
}
\arguments{
\item{x}{x serves as the input: x coordinate}

\item{y}{y serves as the response: y coordinate}
}
\value{
bandwidth of kernel-based local quadratic fit
}
\description{
This function obtains optimal bandwidth to fit a local quadratic regression model \eqn{y = \theta(x)} to the data using a Gaussian kernel and evaluates equilibrium points of the fitted model \eqn{\hat{\theta}^\prime(x)} at the data points.
}
\examples{
set.seed(1234)
x <- runif(1000)
y <- runif(1000)
z <- x^2 + y^2
cond_dens <- estim_conditional(x, y, z, f_eval = 0.25)
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
