% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{fit.all.data}
\alias{fit.all.data}
\title{This function fits optimal CTMC trajectories to timecourse data}
\usage{
fit.all.data(data.inputs, MET.range)
}
\arguments{
\item{data.inputs}{the input datasheet stored in a csv file in the tables dir}

\item{MET.range}{range of time for MET to take place}
}
\value{
the final dataframe with the optimal trajectories fitted to data
}
\description{
This function fits optimal CTMC trajectories to timecourse data
}
