% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_lq}
\alias{fit_lq}
\title{Nonparametric kernel smoothing (NKS)}
\usage{
fit_lq(x, y, h)
}
\arguments{
\item{x}{x serves as the input: x coordinate}

\item{y}{y serves as the response: y coordinate}

\item{h}{h serves as the bandwidth used for local fitting}
}
\value{
values of fitted \eqn{\hat{\theta}(x)} and \eqn{\hat{\theta}^\prime(x)}
}
\description{
This function fits a local quadratic regression model \eqn{y = \theta(x)} to the data using a Gaussian kernel and evaluates equilibrium points of the fitted model \eqn{\hat{\theta}^\prime(x)} at the data points.
}
\examples{
set.seed(1234)
x <- runif(1000)
y <- runif(1000)
z <- x^2 + y^2
cond_dens <- estim_conditional(x, y, z, f_eval = 0.25)
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
