% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{gen_and_spawn}
\alias{gen_and_spawn}
\alias{gen_and_spawn_donors}
\alias{gen_and_spawn_candidates}
\title{Generate bayesian parameters and spawn donors/candidates}
\usage{
gen_and_spawn_donors(desired = "random", days, return_params = FALSE)

gen_and_spawn_candidates(desired = "random", days, return_params = FALSE)
}
\arguments{
\item{desired}{whether the bayesian parameters should be "random" or the "mean"}

\item{days}{number of days (must be greater than > 1)}

\item{return_params}{TRUE or FALSE, to return the paramters}
}
\value{
either a list containing synthetic donors and parameters used to generate their attritubutes or a dataset of synthetic donors
}
\description{
Generate bayesian parameters and spawn donors/candidates
}
\examples{
\dontrun{
d1 <- gen_and_spawn_donors("random", days = 10)
c1 <- gen_and_spawn_candidates("random", days = 10)
}
}
