% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{generate_params}
\alias{generate_params}
\title{Generate Paramters for candidate/donor characterist}
\usage{
generate_params(characteristic, desired = "random")
}
\arguments{
\item{characteristic}{a string specied with "can_" or "don_" and the thecharacteris see details for list of options}

\item{desired}{## either "random" or "mean", a random set of parameters based on a multivariate normal distribution or sample means}
}
\value{
a set of parameters
}
\description{
Generate Paramters for candidate/donor characterist
}
