% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_result.R
\name{getDecisionLog}
\alias{getDecisionLog}
\title{Get the decision log from a COMET result}
\usage{
getDecisionLog(result)
}
\arguments{
\item{result}{A COMETResult object.}
}
\description{
Get the decision log from a COMET result
}
