% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_experiment.R
\name{getResults}
\alias{getResults}
\title{Get the result list from a COMETExperimentResult}
\usage{
getResults(x)
}
\arguments{
\item{x}{A COMETExperimentResult object.}
}
\description{
Get the result list from a COMETExperimentResult
}
