% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_experiment.R
\name{getResultTable}
\alias{getResultTable}
\title{Get the run table from a COMETExperimentResult}
\usage{
getResultTable(x)
}
\arguments{
\item{x}{A COMETExperimentResult object.}
}
\description{
Get the run table from a COMETExperimentResult
}
