% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_experiment.R
\name{getSeeds}
\alias{getSeeds}
\title{Get seeds from a COMETExperimentResult}
\usage{
getSeeds(x)
}
\arguments{
\item{x}{A COMETExperimentResult object.}
}
\description{
Get seeds from a COMETExperimentResult
}
