% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_result.R
\name{getTiming}
\alias{getTiming}
\title{Get timing information from a COMET result}
\usage{
getTiming(result)
}
\arguments{
\item{result}{A COMETResult object.}
}
\description{
Get timing information from a COMET result
}
