% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_removals}
\alias{identify_removals}
\title{Identify}
\usage{
identify_removals(patients, elapsed_time, step_size = 1, cap = NA)
}
\arguments{
\item{patients}{lung transplant candidates}

\item{elapsed_time}{variable for elapsed time}

\item{step_size}{how far back to do conditional survival on (default is 1 day)}

\item{cap}{maximum amount of days to cap at}
}
\value{
candidates removed from waiting list
}
\description{
Identify
}
