% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{iteration}
\alias{iteration}
\title{Iteration}
\usage{
iteration(
  date,
  candidate_database,
  donor_database,
  include_matches,
  updated_list,
  match_alg,
  ...
)
}
\arguments{
\item{date}{current date to simulate}

\item{candidate_database}{full candidate database}

\item{donor_database}{full donor database}

\item{include_matches}{if TRUE returns the donor and candidates matches}

\item{updated_list}{named list of databases that updated at each time point, originally all empty, \cr}

\item{match_alg}{function for matching}

\item{...}{arguments for match algorithm}
}
\value{
a named list of databases that updated at each time point, originally all empty, current_candidates, waitlist_death_database, recipient_database, and post_tx_death_database, non_used_donors
}
\description{
Iteration
}
