% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{KS.label.me}
\alias{KS.label.me}
\title{This function perform Kolmogorov Smirnov scoring on data, credit given to
Priyanka Chakraborty as the code was adapted from her work and modified}
\usage{
KS.label.me(exp.mat, genes, topgenes)
}
\arguments{
\item{exp.mat}{gene expression matrix}

\item{genes}{genes}

\item{top200}{receives top 200 highly variable genes for scoring}
}
\value{
KS score
}
\description{
This function perform Kolmogorov Smirnov scoring on data, credit given to
Priyanka Chakraborty as the code was adapted from her work and modified
}
