% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_policy.R
\name{LASPolicy}
\alias{LASPolicy}
\title{Create a LAS policy}
\usage{
LASPolicy(name = "LAS policy", params = list(), notes = "", autofill = TRUE)
}
\arguments{
\item{name}{Policy name.}

\item{params}{A named list of policy parameters.}

\item{notes}{Optional notes.}

\item{autofill}{Whether to fill missing required keys with NA.}
}
\value{
A LASPolicy object.
}
\description{
Create a LAS policy
}
