% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{match_}
\alias{match_}
\alias{match_las}
\alias{match_cas}
\title{LAS/CAS Matching}
\usage{
match_las(
  cands,
  dons,
  wl_model = NA,
  post_tx_model = NA,
  wl_weight = NA,
  wl_cap = NA,
  post_tx_weight = NA,
  post_tx_cap = NA,
  checks = TRUE
)

match_cas(
  cands,
  dons,
  wl_model = NA,
  post_tx_model = NA,
  wl_weight = NA,
  wl_cap = NA,
  post_tx_weight = NA,
  post_tx_cap = NA,
  bio_weight = NA,
  peds_weight = NA,
  pld_weight = NA,
  efficiency_weight = NA,
  abo_weight = NA,
  height_weight = NA,
  cpra_weight = NA,
  cost_weight = NA,
  distance_weight = NA,
  checks = TRUE
)
}
\arguments{
\item{cands}{candidates of waiting list}

\item{dons}{donors available on given day}

\item{wl_model}{model used for waitlist component of the CAS}

\item{post_tx_model}{model used for post-transplant component of the CAS}

\item{wl_weight}{weight for waitlist survival in days}

\item{wl_cap}{cap for weightlist survival in days}

\item{post_tx_weight}{weight for post-transplant component of CAS}

\item{post_tx_cap}{cap of post-transplatn survival in days}

\item{checks}{whether or not to check the conditions and display warnings, this is there to not check conditions every time the simulation is iterated}

\item{bio_weight}{weight for biological compoent (divided evenly for blood type, hieght, and pra)}

\item{peds_weight}{weight for pediatric patient}

\item{pld_weight}{weight for prior living donor}

\item{efficiency_weight}{weight for travel efficiency}

\item{abo_weight}{weight for blood type compatability (if not specified will default to 1/3 of bio weight)}

\item{height_weight}{weight for height compatability (if not specified will default to 1/3 of bio weight)}

\item{cpra_weight}{weight for cpra compatability (if not specified will default to 1/3 of bio weight)}

\item{cost_weight}{weight given for cost part of CAS (if not specified is half of efficiency weight)}

\item{distance_weight}{weight given for distance part of CAS (if not specified is half of efficiency weight)}
}
\value{
a nested tibble of donors and candidates
}
\description{
LAS/CAS Matching
}
\examples{
match_las(syn_cands, syn_dons, wl_model = "LAS15",
 post_tx_model = "LAS15", wl_weight = 2,
  post_tx_weight = 1, wl_cap = 365, post_tx_cap = 365)
match_cas(syn_cands, syn_dons, wl_model = "CAS23",
 post_tx_model = "CAS23", wl_weight =.25,
  post_tx_weight = .25, wl_cap = 365, post_tx_cap = 1826,
 bio_weight = .15, pld_weight = 0.05,
 peds_weight = 0.2, efficiency_weight = 0.1)
}
