% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_survival}
\alias{mean_survival}
\title{Mean Survival}
\usage{
mean_survival(lp, surv_rates)
}
\arguments{
\item{lp}{vector of linear predictors}

\item{surv_rates}{vector of survival times}
}
\description{
This function calculates mean survival
}
\examples{
lps <- c(0, 1, -.5)
surv_rates <- seq(1,.85, -.0005)
mean_survival(lps, surv_rates)
}
