% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacif.R
\name{metacinc}
\alias{metacinc}
\title{Meta-analysis of CIFs (ratios)}
\usage{
metacinc(event, N, ptime, cause, time, group, study, control,
  predt = NULL, data = NULL, subset = NULL, pool = FALSE,
  landmark = FALSE, plot = FALSE, binn = 5)
}
\description{

}
\details{
the function has print and plot methods.
}
\seealso{
calccinc
}
