% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_result.R
\name{metrics}
\alias{metrics}
\title{Extract comparison metrics from a COMET result}
\usage{
metrics(result)
}
\arguments{
\item{result}{A COMETResult object.}
}
\description{
Extract comparison metrics from a COMET result
}
