% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_las.R
\name{rmst}
\alias{rmst}
\alias{rmst.character}
\title{Restricted Meant Survival Time}
\usage{
rmst(object, ...)

\method{rmst}{character}(object, cand_data, cap = NA, wl = TRUE, ...)
}
\arguments{
\item{object}{only character objects of 'LAS15', 'LAS21' and 'CAS23' are acceptable}

\item{...}{arguement given from other methods}

\item{cand_data}{candidate data}

\item{cap}{cap for truncation survival}

\item{wl}{TRUE (waitlist) or FALSE (post_transplant)}
}
\value{
a dateset of candidates, linear predictor and restricted mean survival time in days
}
\description{
Restricted Meant Survival Time
}
\examples{
rmst("LAS21", syn_cands, 365, TRUE)
}
