% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{rsamp}
\alias{rsamp}
\title{Resample Function}
\usage{
rsamp(FUN, limits, ...)
}
\arguments{
\item{FUN}{sample function to use}

\item{limits}{limits of values set as a vector c(x,y), Inf and -Inf are acceptable values}

\item{...}{other arguments for FUN}
}
\value{
a vector a randomly sampled values within the limits
}
\description{
Resample Function
}
\examples{
y <- rsamp(rnorm, limits = c(0,5), n = 100, mean = 0, mean = 3)
y <- rsamp(rnorm, limits = c(0, Inf), n = 100, mean = 0, mean = 3)
}
