% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{run.CTMC}
\alias{run.CTMC}
\title{This function generates trajectories for a CTMC model given parameters}
\usage{
run.CTMC(alph_fun, time.range, M_sc_fun, Mu_sc_fun, p0_fun)
}
\arguments{
\item{alph_fun}{alpha parameter used in function}

\item{time.range}{specific timerange to generate trajectory}

\item{M_sc_fun}{ratio of M/E at steady state}

\item{Mu_sc_fun}{ratio of H/E at steady state}

\item{p0_fun}{initial state vector}
}
\value{
trajectories along with the resulting p vector
}
\description{
This function generates trajectories for a CTMC model given parameters
}
