% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{run_pipeline}
\alias{run_pipeline}
\title{This function runs the data driven pipeline for inferring trajectories}
\usage{
run_pipeline(data.inputs, tables.dir, input.data.dir, cutoff)
}
\arguments{
\item{data.inputs}{the input datasheet stored in a csv file in the tables dir}

\item{tables.dir}{directory with a csv file that had the address
for the data and metadata}

\item{input.data.dir}{this is the directory where the input file
should be saved in}

\item{cutoff}{of highly variable EMT genes to be considered}
}
\value{
the inferred trajectories, also saves them within the
COMET_populated_files directory
}
\description{
This function runs the data driven pipeline for inferring trajectories
}
