% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run Simulation}
\usage{
run_simulation(
  days,
  can_start = 1250,
  match_alg = match_cas(),
  ...,
  seed = NULL,
  desired = "random",
  return_params = FALSE,
  include_matches = FALSE
)
}
\arguments{
\item{days}{number of days to simulate}

\item{can_start}{number of candidates to start on the waitlist, they will be randomly selected to be more reflective of a waiting period listing day, all of these candidates will have a negative c_id and listing_day}

\item{match_alg}{function of how to screen and match individuals}

\item{...}{arguments for match_alg}

\item{seed}{seed to set}

\item{desired}{random or mean for parameter generation}

\item{return_params}{if TRUE returns the parameters used for generating donors and candidates}

\item{include_matches}{if TRUE returns the donor and candidates matches}
}
\value{
list of full donors and candidates simulated at beginning and final datasets, who's left on waiting list, transplanted, died on waiting list, died post transplant and non utililized donors
}
\description{
Run Simulation
}
\examples{
\dontrun{
# Please ensure that data package is downloaded before running this example.
# If it is not, see the ‘COMET’ package readme for details
# on installing the required data package.
if (requireNamespace("cometdata", quietly = TRUE)) {
r1 <- run_simulation(days = 20, can_start = 1000,
 match_alg = match_las, wl_model = "LAS15", post_tx_model = "LAS15",
 wl_weight = 1, post_tx_weight = 1, wl_cap = 365, post_tx_cap = 365)
r2 <- run_simulation(days = 20, can_start = 1000,
 match_alg = match_cas, wl_model = "CAS23", post_tx_model = "CAS23",
 wl_weight = 0.25, post_tx_weight = 0.25, wl_cap = 365,
 post_tx_cap = 1826, bio_weight = .15, pld_weight = 0.05,
 peds_weight = 0.2, efficiency_weight = 0.1)
 }
 }
}
