% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_experiment.R
\name{runExperiment}
\alias{runExperiment}
\title{Run a COMET experiment}
\usage{
runExperiment(experiment)
}
\arguments{
\item{experiment}{A COMETExperiment object.}
}
\description{
Run a COMET experiment
}
