% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS.R
\name{scale_pre_post}
\alias{scale_pre_post}
\title{Pre and Post Survival Scaling}
\usage{
scale_pre_post(
  data,
  pre,
  post,
  pre_weight = NA,
  post_weight = NA,
  pre_cap = NA,
  post_cap = NA
)
}
\arguments{
\item{data}{set containing waitlist and post expected days survived}

\item{pre}{variable for waitlist expected survival}

\item{post}{variable for post transplant expected survival}

\item{pre_weight}{how much you wanted to weight waitlist (default is 1)}

\item{post_weight}{how much you want to weight post-transplant (default is 1)}

\item{pre_cap}{cap on waitlist survival}

\item{post_cap}{cap on post-transplant survival}
}
\value{
a dataset with the added variable ly_score which is a score 0-1.
}
\description{
Pre and Post Survival Scaling
}
