% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen}
\alias{screen}
\alias{height_screen}
\alias{abo_screen}
\alias{pra_screen}
\alias{count_screen}
\alias{las_dist_calc}
\alias{dist_calc}
\alias{las_offer_rank}
\alias{cas_offer_rank}
\title{Screen Candidates and Donors for Compatibility}
\usage{
height_screen(cands, dons)

abo_screen(cands, dons)

pra_screen(cands, dons)

count_screen(cands, dons)

las_dist_calc(cands, dons)

dist_calc(cands, dons)

las_offer_rank(..., overall_ranking)

cas_offer_rank(
  ...,
  overall_ranking,
  efficiency_weight = 0.1,
  cost_weight = NA,
  distance_weight = NA,
  checks = TRUE
)
}
\arguments{
\item{cands}{tibble of candidates}

\item{dons}{tibble of donors}

\item{...}{tibbles of screening conditions, height, blood, pra, distance, etc}

\item{overall_ranking}{tibble containing overall ranking by candidate}

\item{efficiency_weight}{weight of efficiency for distance between candidate and donor}

\item{cost_weight}{weight given for cost part of CAS (if NA specified is half of efficiency weight)}

\item{distance_weight}{weight given for distance part of CAS (if NA specified is half of efficiency weight)}

\item{checks}{whether or not to check the conditions and display warnings, this is there to not check conditions every day when the simulation is iterated}
}
\value{
a dataset containing compatible donors and candidates based on specific criteria or a dataset of all comprable matches and their respective rankings
}
\description{
Screen for compatible candidates and donors based on biological factors
}
\details{
\code{height_screen} screens compatible donors and candidates based on acceptable height ranges.

\code{abo_screen} screens donors and compatible donors based on blood type

\code{pra_screen} not yet implemented will screen on pra compatibility organs.

\code{count_screen} Makes sure double lung transplant are matched with double lung donors, singles can be matched and received a one of the double lung

\code{las_dist_calc} Returns the distance between the matched candidates and donors with LAS distance categories.

Matching rules for LAS distances are broken into the following zones, 250 nautical miles, 500, 1000, 1500, 2500, the US
}
\examples{
## The functions will use the internal cands and dons stored in the package
height_match <- height_screen(syn_cands, syn_dons)
abo_match <- abo_screen(syn_cands, syn_dons)
pra_match <- pra_screen(syn_cands, syn_dons)
count_match <- count_screen(syn_cands, syn_dons)
las_dists <- las_dist_calc(syn_cands, syn_dons)
cas_dists <- dist_calc(syn_cands, syn_dons)
las_rank <- calculate_las(syn_cands, wl_model = "LAS15", post_tx_model = "LAS15",
 wl_cap = 365, post_tx_cap = 365, wl_weight = 2, post_tx_weight = 1)
las_offers <- las_offer_rank(height_match, abo_match, count_match, pra_match, las_dists,
 overall_ranking = las_rank)
cas_rank <- calculate_sub_cas(syn_cands, wl_model = "CAS23", post_tx_model = "CAS23",
 wl_weight = 0.25, post_tx_weight = 0.25, bio_weight = 0.15, pld_weight = 0.05, peds_weight = 0.2)
cas_offers <- cas_offer_rank(height_match, abo_match, count_match, pra_match, cas_dists,
 overall_ranking = cas_rank)
}
