% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{spec_day}
\alias{spec_day}
\title{Grab those alive on specific day on the waitlist}
\usage{
spec_day(COMET, day)
}
\arguments{
\item{COMET}{a COMET object}

\item{day}{a number}
}
\value{
a tibble of candidates on waitling list at a specific day
}
\description{
Grab those alive on specific day on the waitlist
}
\examples{
## will add later
}
