% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{start.comet}
\alias{start.comet}
\title{Read necessary files and parameters for pipeline to run
This function reads all the necessary files to run this pipeline}
\usage{
\method{start}{comet}(tables.dir, input.data.dir, input.files.dir)
}
\arguments{
\item{tables.dir}{directory with a csv file that had the address
for the data and metadata}

\item{input.data.dir}{this is the directory where the input file
should be saved in}

\item{input.files.dir}{this directory has all the necessary params
and files to run the pipeline}
}
\value{
nothing, stores necessary parameters for the model in global
variables
}
\description{
Read necessary files and parameters for pipeline to run
This function reads all the necessary files to run this pipeline
}
