% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_loop_c}
\alias{sum_loop_c}
\title{Sum loop C}
\usage{
sum_loop_c(x, y)
}
\arguments{
\item{x}{vector of linear predictors}

\item{y}{vector of survival times}
}
\description{
This function loops over the calculate mean survival
}
