% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET_functions.R
\name{transition_matrix}
\alias{transition_matrix}
\title{This function finds the transition matrix given the parameters for the
generator matrix}
\usage{
transition_matrix(lambda_E, lambda_M, mu_E, mu_M, t)
}
\arguments{
\item{lambda_E}{Rate of transition from H to E}

\item{lambda_M}{Rate of transition from H to M}

\item{mu_E}{Rate of transition from E to H}

\item{mu_M}{Rate of transition from M to H}

\item{t}{time}
}
\value{
the probability transition matrix
}
\description{
This function finds the transition matrix given the parameters for the
generator matrix
}
