% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transplant.R
\name{transplant_candidates}
\alias{transplant_candidates}
\title{Transplant Candidates}
\usage{
transplant_candidates(matches, rec_ids, max_offer = NA)
}
\arguments{
\item{matches}{nested tibble of matches}

\item{rec_ids}{vector of c_id of transplanted candidates}

\item{max_offer}{maximum number of offers to consider}
}
\value{
a dataset of candidates that have been transplanted with their respective donors
}
\description{
Transplant Candidates
}
\examples{
## Will add later
}
