% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_las.R
\name{trunc_days.character}
\alias{trunc_days.character}
\title{Title}
\usage{
\method{trunc_days}{character}(object, cand_data, cap = NA, wl = TRUE)
}
\arguments{
\item{wl}{TRUE (waitlist) or FALSE (post_transplant)}
}
\description{
Title
}
