% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_las.R
\name{trunc_days}
\alias{trunc_days}
\title{Truncated Survival days}
\usage{
trunc_days(object, ...)
}
\arguments{
\item{object}{survival object (only coxph models currently are accepted)}

\item{cand_data}{candidate data}

\item{cap}{cap for truncation survival}
}
\description{
Truncated Survival days
}
