% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{update}
\alias{update}
\alias{update_patients}
\alias{identify_removals}
\alias{identify_deaths}
\title{Update Patients}
\usage{
update_patients(
  patients,
  model = NULL,
  elapsed_time,
  step_size,
  pre_tx = TRUE,
  cap = NA,
  date
)

identify_removals(patients, elapsed_time, step_size = 1, cap = NA)

identify_deaths(
  patients,
  model = NULL,
  elapsed_time,
  step_size,
  pre_tx = TRUE,
  cap = NA,
  date
)
}
\arguments{
\item{patients}{dataset of lung transplant patients}

\item{model}{which model to use for identifying deaths}

\item{elapsed_time}{name of variable for elasped time}

\item{step_size}{how far back for conditional survival (default is 1 day)}

\item{pre_tx}{is deaths pre or post transplant}

\item{cap}{cap for suvival estiamte}

\item{date}{date}
}
\value{
A list of candidates who were Alive, Dead, new_char for new characteristics and removed if on waiting list

a dataset of candidates and whether or not they were removed from the waiting list

a dataset of candidates and whether or not they experienced death on the waitlist
}
\description{
Update Patients
}
\examples{
identify_deaths(cands, "CAS23", days_on_waitlist, 1, TRUE, 365, 16)
identify_removals(cands, days_on_waitlist, 1, TRUE, 2365, 16)
update_patients(cands, "CAS23r", days_on_waitlist, 1, TRUE, 365, 16)
}
