% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{update_patients}
\alias{update_patients}
\title{Title}
\usage{
update_patients(
  patients,
  model = NULL,
  elapsed_time,
  step_size,
  pre_tx = TRUE,
  cap = NA,
  date
)
}
\arguments{
\item{patients}{dataset of lung transplant patients}

\item{model}{which model to use for identifying deaths}

\item{elapsed_time}{name of variable for elasped time}

\item{step_size}{how far back for conditional survival (default is 1 day)}

\item{pre_tx}{is deaths pre or post transplant}

\item{cap}{cap for suvival estiamte}

\item{date}{date}
}
\description{
Title
}
