% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_config.R
\name{validateConfig}
\alias{validateConfig}
\title{Validate a COMET configuration}
\usage{
validateConfig(config)
}
\arguments{
\item{config}{A COMETConfig object.}
}
\description{
Validate a COMET configuration
}
