% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_experiment.R
\name{validateExperiment}
\alias{validateExperiment}
\title{Validate a COMET experiment}
\usage{
validateExperiment(experiment)
}
\arguments{
\item{experiment}{A COMETExperiment object.}
}
\description{
Validate a COMET experiment
}
