% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_simulator.R
\name{validateSimulator}
\alias{validateSimulator}
\title{Validate a COMET simulator}
\usage{
validateSimulator(simulator)
}
\arguments{
\item{simulator}{A COMETSimulator object.}
}
\description{
Validate a COMET simulator
}
