% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_organ.R
\name{acceptance_prob}
\alias{acceptance_prob}
\title{Acceptance Probability from matched data}
\usage{
acceptance_prob(matched_data, dons, cands)
}
\arguments{
\item{matched_data}{output from matching module}

\item{dons}{generated donor data}

\item{cands}{generated candidate data}
}
\value{
matched_data with acceptance probabilities
}
\description{
Acceptance Probability from matched data
}
\examples{
dplyr::mutate(syn_matches, offer_rank = rank(-lu_score), .by = d_id) |>
 acceptance_prob(dons = syn_dons, cands = syn_cands)
}
