% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cas.R
\name{calculate_cas_dist}
\alias{calculate_cas_dist}
\title{Calculate continuous distribution}
\usage{
calculate_cas_dist(
  match_data,
  efficiency_weight = NA,
  cost_weight = NA,
  distance_weight = NA
)
}
\arguments{
\item{match_data}{dataset of patients containing covaraites for las calculations and other continuous distribution factors}

\item{efficienty_weight}{weight given for efficiency (cost and distance)}
}
\value{
a dataset complete with the efficient/distance metric
}
\description{
Calculate continuous distribution
}
\examples{
## will add real examples later
## This needs to be thought out, only the distance changes with the donors
}
