% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS.R
\name{calculate_las}
\alias{calculate_las}
\title{Calculate LAS scores}
\usage{
calculate_las(
  wl_data,
  wl_model = NULL,
  post_tx_model = NULL,
  wl_cap = NA,
  post_tx_cap = NA,
  wl_weight = NA,
  post_tx_weight = NA,
  checks = TRUE
)
}
\arguments{
\item{wl_data}{data to calculate las}

\item{wl_model}{which waitlist model to use.
Options are "LAS15","LAS21", "CAS23"}

\item{post_tx_model}{which post transplant model to use.
Options are "LAS15","LAS21", "CAS23"}

\item{wl_cap}{cap of waitlist survival}

\item{post_tx_cap}{cap of post-transplant survival}

\item{wl_weight}{relative weight of waitlist mortality}

\item{post_tx_weight}{relative weight of post-transplant mortality}

\item{checks}{whether or not to check the conditions and display warnings, this is there to not check conditions every day the simulation is iterated}
}
\value{
a dataset of candidates and their LAS score on a 0 to 1 scale and linear predictors for waitlist and post-transplant components.
To get LAS/CAS score simply multiply by 100.
}
\description{
Calculate LAS scores
}
\examples{
calculate_las(syn_cands, wl_model = "LAS21", post_tx_model = "LAS21",
 wl_cap = 365, post_tx_cap = 365, wl_weight = 2, post_tx_weight = 1)
}
