% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET.R
\docType{package}
\name{COMET}
\alias{COMET}
\title{COMET}
\section{Initialization components}{

\itemize{
\item \code{\link{gen_and_spawn_candidates}}
\item \code{\link{gen_and_spawn_donors}}
}
}

\section{Iterative components}{

\itemize{
\item \code{\link{transplant_candidates}}
\item \code{\link{update_patients}}
\item \code{\link{match_cas}}
}
}

\section{Families}{

In addition to the partition of the simulation into initialization and iterative aspects,
functions generally track either only candidates, only donors, or both candidates and donors, so
allow for package description in this manner.  This allows for families of functions to be classified by the
primary agents they manipulate.
}

