% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMET.R
\docType{package}
\name{COMET}
\alias{COMET}
\title{COMET-package}
\description{
The COMET package provides a open source framework to simulate the lung allocation system of the United States.
}
\section{Funding}{

This project was funded by the National Heart, Lung, and Blood Institute of the National Institutes of Health (R01HL153175). The content is solely the responsibility of the authors and does not necessarily represent the official views of the National Institutes of Health.
}

