% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{spawn_donors}
\alias{spawn_donors}
\alias{gen_and_spawn_donors}
\alias{spawn_candidates}
\alias{gen_and_spawn_candidates}
\title{Generate bayesian parameters and spawn donors/candidates}
\usage{
spawn_donors(
  days,
  daily_lambdas_center,
  sex_thetas,
  race_mat_beta,
  hgt_params,
  abo_mat_beta,
  age_params,
  cod_mat_beta,
  smoke_hist_mat_beta,
  don_org_odds,
  dcd_params,
  don_util_odds
)

gen_and_spawn_donors(desired = "random", days, return_params = FALSE)

spawn_candidates(
  days,
  daily_lambdas_center,
  sex_thetas,
  race_mat_beta,
  abo_mat_beta,
  age_params,
  hgt_params,
  wgt_params,
  wgt_params_c,
  diab_params,
  airway_params,
  oxygen_params,
  resp_mat_beta,
  surg_mat_beta,
  fev_params,
  fvc_params,
  pco2_params,
  pf_params,
  po2_params,
  mpap_params,
  o2_freq_mat,
  vent_mat_beta,
  supp_o2_params,
  walk6m_params,
  bili_params,
  creat_params,
  spap_params,
  ci_params,
  cvp_params,
  fs_mat_beta,
  pco2_15_params,
  sa_params,
  lam_params,
  eisen_params,
  sd_params,
  ar_mat_beta,
  dr_mat_beta
)

gen_and_spawn_candidates(desired = "random", days, return_params = FALSE)
}
\arguments{
\item{days}{number of days to simulate over}

\item{daily_lambdas_center}{lambda per candidate center}

\item{sex_thetas}{theta paramter to assign sex}

\item{race_mat_beta}{beta matrix for race multinomial}

\item{hgt_params}{height}

\item{abo_mat_beta}{blood type}

\item{age_params}{age}

\item{cod_mat_beta}{cause of death}

\item{smoke_hist_mat_beta}{smoking history}

\item{don_org_odds}{left, right or double lung}

\item{dcd_params}{dcd status}

\item{don_util_odds}{utilization rate (transplanted in US)}

\item{desired}{"random" or "mean"}

\item{return_params}{TRUE or FALSE, number of paramters, ouptut will be a list of a dataset and the parameters}

\item{wgt_params}{weight}

\item{wgt_params_c}{weight diagnosis for Group C}

\item{diab_params}{diabetes}

\item{airway_params}{airway function}

\item{oxygen_params}{oxygen function}

\item{resp_mat_beta}{respiratory support}

\item{surg_mat_beta}{surgery needed}

\item{fev_params}{fev1}

\item{fvc_params}{fvc}

\item{pco2_params}{pco2}

\item{pf_params}{pf}

\item{po2_params}{po2}

\item{mpap_params}{mean pulmonary artery pressure}

\item{o2_freq_mat}{oxygen frequency}

\item{vent_mat_beta}{ventilator status}

\item{supp_o2_params}{supplemental oxygen}

\item{walk6m_params}{six minute walk distance}

\item{bili_params}{bilirubin}

\item{creat_params}{creatinine}

\item{spap_params}{systolic pulmonary artery pressue}

\item{ci_params}{cardiac index}

\item{cvp_params}{central venous pressure}

\item{fs_mat_beta}{functional status}

\item{pco2_15_params}{pco2 threshold}

\item{sa_params}{sarcoidosis Group A}

\item{lam_params}{lam}

\item{eisen_params}{eisenmengers}

\item{sd_params}{sarcoidosis Group D}

\item{ar_mat_beta}{other Group A, diseases}

\item{dr_mat_beta}{other Group D, disesase}

\item{dx_mat_beta}{diangosis group}
}
\value{
dataset of donors

either a list containing synthetic donors and parameters used to generate their attritubutes or a dataset of synthetic donors

dataset of synthetic candidates over a set number of days
}
\description{
Generate bayesian parameters and spawn donors/candidates
}
\examples{
d1 <- gen_and_spawn_donors("random", days = 10)
c1 <- gen_and_spawn_candidates("random", days = 10)
}
