% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{gen_and_spawn}
\alias{gen_and_spawn}
\alias{gen_and_spawn_donors}
\alias{gen_and_spawn_candidates}
\title{Generate bayesian parameters and spawn donors/candidates}
\usage{
gen_and_spawn_donors(desired = "random", days, return_params = FALSE)

gen_and_spawn_candidates(desired = "random", days, return_params = FALSE)
}
\arguments{
\item{desired}{whether the Bayesian parameters should be "random" or the "mean"}

\item{days}{number of days (must be greater than > 1)}

\item{return_params}{if TRUE returns the parameters used for generating donors and candidates}
}
\value{
if return_params = TRUE a list containing synthetic donors or candidates and parameters used to generate their attributes
 Otherwise a data set of synthetic donors or candidates
}
\description{
Generate bayesian parameters and spawn donors/candidates
}
\examples{
\dontrun{
d1 <- gen_and_spawn_donors("random", days = 10)
c1 <- gen_and_spawn_candidates("random", days = 10)
}
}
