% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_las.R
\name{get_surv_data}
\alias{get_surv_data}
\alias{get_surv_data.survfit}
\alias{get_surv_data.coxph}
\title{Get Survival Data from coxph and survfit object}
\usage{
get_surv_data(object, ...)

\method{get_surv_data}{survfit}(object, ...)

\method{get_surv_data}{coxph}(object, ...)
}
\arguments{
\item{object}{either a survfit or coxph object}

\item{...}{other arguments depending on class}
}
\value{
a data.frame of survival times from a survfit of coxph
}
\description{
Get Survival Data from coxph and survfit object
}
\examples{
# Create the simplest test data set
## same example from survival package
test1 <- list(time=c(4,3,1,1,2,2,3),
            status=c(1,1,1,0,1,1,0),
           x=c(0,2,1,1,1,0,0),
           sex=c(0,0,0,0,1,1,1))
# Fit a stratified model
c1 <- survival::coxph(survival::Surv(time, status) ~ x + survival::
strata(sex), test1)
km1 <- survival::survfit(survival::Surv(time, status) ~ 1, test1)
get_surv_data(c1)
get_surv_data(km1)
}
