% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{iteration}
\alias{iteration}
\title{Iteration}
\usage{
iteration(
  date,
  candidate_database,
  donor_database,
  include_matches,
  updated_list,
  match_alg,
  ...
)
}
\arguments{
\item{date}{current date to simulate}

\item{candidate_database}{full candidate database, from initiialization}

\item{donor_database}{full candidate database, from initiialization}

\item{include_matches}{TRUE or FALSE, whether or not the include the donor and candidates matches
current_candidates, waitlist_death_database, recipient_database, and post_tx_death_database, discarded_donors}

\item{updated_list}{named list of databases that updated at each time point, originally all empty, \cr}

\item{match_alg}{function for matching}

\item{...}{arguemnts for match algorithm}
}
\value{
a named list of databases that updated at each time point, originally all empty, current_candidates, waitlist_death_database, recipient_database, and post_tx_death_database, discarded_donors
}
\description{
Iteration
}
