% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run Simulation}
\usage{
run_simulation(
  days,
  can_start = 1250,
  waitlist_update_freq = 1,
  post_tx_update_freq = 1,
  match_alg = match_cas(),
  extra_days = 0,
  ...,
  seed = NULL,
  desired = "random",
  return_params = FALSE,
  include_matches = FALSE
)
}
\arguments{
\item{days}{number of days to simulate}

\item{can_start}{number of candidates to start on the waitlist, they will be randomly selected to be mor listing day, all of these candidates will have a negative c_id and listing_day}

\item{waitlist_update_freq}{how often to update waitlist patients}

\item{post_tx_update_freq}{how often to update post_tx patients}

\item{match_alg}{function of how to screen and match individuals}

\item{extra_days}{how many days to simulate deaths}

\item{...}{arguments for match_alg}

\item{seed}{seed to set}

\item{desired}{random or mean for parameter generation}

\item{return_params}{TRUE or FALSE see the paramters used for generating donors and candidates}

\item{include_matches}{TRUE or FALSE, whether or not the include the donor and candidates matches}
}
\value{
list of full donors and candidates simluated at beginning and final datasets, who's left on waitlist, transplanted, death on both and discarded donors
}
\description{
Run Simulation
}
\examples{

r1 <- run_simulation2(days = 20, can_start = 1000, waitlist_update_freq = 1, post_tx_update_freq = 1, match_alg = match_las, wl_model = "LAS15", post_tx_model = "LAS15", wl_weight = 1, post_tx_weight = 1, wl_cap = 365, post_tx_cap = 365)
r2 <- run_simulation2(days = 20, can_start = 1000, waitlist_update_freq = 1, post_tx_update_freq = 1, match_alg = match_cas, wl_model = "CAS23", post_tx_model = "CAS23", wl_weight = 0.25, post_tx_weight = 0.25, wl_cap = 365, post_tx_cap = 1825, bio_weight = .15, pld_weight = 0.05, peds_weight = 0.2, efficiency_weight = 0.1)
}
