% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_las.R
\name{trunc_days}
\alias{trunc_days}
\alias{trunc_days.character}
\title{Truncated Survival days}
\usage{
trunc_days(object, ...)

\method{trunc_days}{character}(object, cand_data, cap = NA, wl = TRUE, ...)
}
\arguments{
\item{object}{only character objects of 'LAS15', 'LAS21' and 'CAS23' are acceptable}

\item{...}{arguement given from other methods}

\item{wl}{TRUE (waitlist) or FALSE (post_transplant)}
}
\value{
a dateset of candidates, linear predictor and mean survival
}
\description{
Truncated Survival days
}
\examples{
trunc_days("LAS21", syn_cands, 365, TRUE)
}
