% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{update}
\alias{update}
\alias{update_patients}
\alias{identify_removals}
\alias{identify_deaths}
\title{Update Patients}
\usage{
update_patients(
  patients,
  model = NULL,
  elapsed_time,
  pre_tx = TRUE,
  cap = NA,
  date
)

identify_removals(patients, elapsed_time, cap = NA)

identify_deaths(
  patients,
  model = NULL,
  elapsed_time,
  pre_tx = TRUE,
  cap = NA,
  date
)
}
\arguments{
\item{patients}{dataset of lung transplant patients}

\item{model}{which model to use for identifying deaths}

\item{elapsed_time}{name of variable for elapsed time}

\item{pre_tx}{is the updating for pre or post transplant}

\item{cap}{cap for survival calculation}

\item{date}{current date (numeric)}
}
\value{
\code{update_patients} returns a list of candidates who were Alive, Dead, new_char for new characteristics and removed if on waiting list

\code{identify_removals} returns a dataset of candidates and whether or not they were removed from the waiting list (not for being too sick)

\code{identify_deaths} returns a dataset of candidates and whether or not they experienced death or were too sick to transplant while on the waiting list
}
\description{
Updating patients status and characteristics
}
\examples{
identify_deaths(syn_cands, model = "CAS23", elapsed_time = days_on_waitlist,
   pre_tx = TRUE, cap = 365,  date = 16)
identify_removals(syn_cands, elapsed_time =  days_on_waitlist, cap = 2365)
update_patients(syn_cands, model = "CAS23r", elapsed_time =days_on_waitlist,
  pre_tx = TRUE, cap = 365, date = 16)
}
